move toml config to sodar api
